refactor(SimpleFeederSubsystem): Update to new organization scheme
